dataset_scripts/requirements.txt
python-dotenv
PyMySQL
psycopg2
pandas
requests
pytest
openpyxl
#gql[all]
colorist
sentence-transformers
